+++ CHANGELOG.md
@@ -9,8 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Add an additional boolean parameter to `XpringClient`'s constructor which allows toggling between protocol buffer implementations.
-- Switch the protocol buffer generator to use gRPC-Web instead of gRPC-Node when on browser (or when forcing HTTP) for cross environment compatability with web and node.
-- Add XMLHttpRequest to the global namespace for compatability with gRPC-Web when forcing HTTP on node.
+- Switch BigInt to use the big-integer polyfill lib.
 
 ### Fixed
 - Implement `XpringClient`'s `getBalance` method using rippled's protocol buffers.